Moved out tests for illumina_finished_msg.py in to their own file!
